Move category and utils classes
